+++ Eradication/IndividualTyphoid.h
@@ -90,7 +90,7 @@ namespace Kernel
         virtual void UpdateInfectiousness(float dt);
         virtual void Update(float currenttime, float dt);
         virtual Infection* createInfection(suids::suid _suid);
-        virtual void AcquireNewInfection( const IStrainIdentity *infstrain = nullptr, int incubation_period_override = -1) override;
+        virtual void AcquireNewInfection( const IStrainIdentity *infstrain = nullptr, float incubation_period_override = -1.0f ) override;
         virtual float GetImmunityReducedAcquire() const override;
         virtual HumanStateChange GetStateChange() const;
         //virtual void applyNewInterventionEffects(float dt);
